Handle the case where token is not fetched before attempting to edit.
Since [February 2020](https://lists.wikimedia.org/pipermail/mediawiki-api-announce/2020-February/000151.html),
absence of the token parameter in the query does *not* give a 'notoken' error, rather it gives the standard
'missingparams' error which may indicate absence of any required parameter.

By setting the initial editToken as a garbage value (`%notoken%`), we ensure that no query is sent without the
token. If this garbage value token is sent, a badtoken error results, which we then handle (previous commit).