Overhaul logging interface
Remove semlog dependency which kept track of statistics and retained logged messages in memory (as "log history").

Semlog is modified and restructured to log.ts. This also removes the use of `global` scoped variables and makes it a TypeScript-style import.

The issue of mwn.log being static but the constructor being able to set logging configs is also sorted.