Use action=query instead of action=parse for methods in page class
Breaking change for v3. The query action is much more performant as it reads from the db instead of triggering the parser. This does mean that only the first 50/500 results are returned, which seems reasonable as continuedQuery()/continuedQueryGen() is available for use cases that need full results. Category sortkeys and red/blue status of links are no longer available in the results.